Change 2025 Day 6 to Long type solution
Added example and main input, and example test value for Part 1

+++ src/main/java/site/ayrilea/advent/solution/year2025/day6/Part1.java
@@ -5,10 +5,10 @@ import site.ayrilea.advent.solution.Solution;
 import site.ayrilea.advent.solution.SolutionMetadata;
 
 @SolutionMetadata(year = 2025, day = 6, part = 1)
-public class Part1 implements Solution<Integer> {
+public class Part1 implements Solution<Long> {
 
     @Override
-    public Integer solve(Input input) {
-        return 0;
+    public Long solve(Input input) {
+        return 0L;
     }
 }
